invoice created

+++ src/api/models/actions.py
@@ -1,5 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import Integer, String, DateTime,ForeignKey
+from sqlalchemy import Integer, String, DateTime,ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime
 from .database import db
@@ -18,7 +18,7 @@ class Action(db.Model):
     start_date: Mapped[datetime] = mapped_column(DateTime, nullable=False)
     description: Mapped[str] = mapped_column(String(350), nullable=False)
     contractor: Mapped[str] = mapped_column(String(255), nullable=False)
-    bill_amount: Mapped[int] = mapped_column(Integer, nullable=True)
+    bill_amount: Mapped[float] = mapped_column(Numeric(10,2), nullable=True)
     bill_image: Mapped[str] = mapped_column(String(255), nullable=True)
     issue_id: Mapped[int] = mapped_column(ForeignKey('issues.id'), nullable=False)
     issue: Mapped ['Issue'] = relationship(
